@@ -1,116 +1,41 @@
 {
-  "name": "react-redux-starter-kit",
-  "version": "1.0.0",
-  "description": "Get started with React, Redux, and React-Router!",
-  "main": "index.js",
-  "engines": {
-    "node": ">=4.2.0",
-    "npm": "^3.0.0"
-  },
+  "private": true,
   "scripts": {
-    "clean": "rimraf dist",
-    "compile": "babel-node bin/compile",
-    "lint": "eslint .",
-    "lint:fix": "npm run lint -- --fix",
-    "start": "better-npm-run start",
-    "dev": "better-npm-run dev",
-    "dev:nw": "npm run dev -- --nw",
-    "dev:no-debug": "npm run dev -- --no_debug",
-    "test": "echo 'No tests so far'",
-    "test:dev": "npm run test -- --watch",
-    "deploy": "better-npm-run deploy",
-    "flow:check": "babel-node bin/flow-check",
-    "codecov": "cat coverage/*/lcov.info | codecov"
+    "build": "webpack",
+    "lint": "eslint src",
+    "start": "webpack-dev-server -d --hot --inline --progress --colors --history-api-fallback --host 0.0.0.0 --port 4000",
+    "test": "npm run lint && npm run check",
+    "deploy": "rm -rf dist && npm run compile",
+    "check": "babel-node bin/flow-check"
   },
-  "betterScripts": {
-    "dev": {
-      "command": "nodemon --exec babel-node bin/server",
-      "env": {
-        "NODE_ENV": "development",
-        "DEBUG": "app:*"
-      }
-    },
-    "deploy": {
-      "command": "npm run clean && npm run compile",
-      "env": {
-        "NODE_ENV": "production",
-        "DEBUG": "app:*"
-      }
-    },
-    "start": {
-      "command": "babel-node bin/server",
-      "env": {
-        "DEBUG": "app:*"
-      }
-    },
-    "test": {
-      "command": "babel-node ./node_modules/karma/bin/karma start bin/karma",
-      "env": {
-        "NODE_ENV": "test",
-        "DEBUG": "app:*"
-      }
-    }
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/davezuko/react-redux-starter-kit.git"
-  },
-  "author": "David Zukowski <david@zuko.me> (http://zuko.me)",
-  "license": "MIT",
   "metadata": {
     "graphql": {
-      "schema": "./schema.json"
+      "schema": "./.tmp/schema.json"
     }
   },
   "dependencies": {
-    "better-npm-run": "0.0.8",
-    "bulma": "0.0.16",
-    "co-request": "^1.0.0",
-    "debug": "^2.2.0",
-    "font-awesome": "^4.5.0",
-    "history": "^2.0.1",
-    "iconv-lite": "^0.4.13",
-    "immutable": "^3.7.6",
-    "koa": "^2.0.0-alpha.3",
-    "koa-connect-history-api-fallback": "^0.3.0",
-    "koa-convert": "^1.2.0",
-    "koa-static": "^2.0.0",
+    "normalize.css": "^3.0.3",
     "react": "^0.14.7",
     "react-addons-pure-render-mixin": "^0.14.7",
     "react-dom": "^0.14.7",
-    "react-redux": "^4.4.0",
     "react-relay": "^0.7.3",
     "react-router": "^2.0.0",
-    "react-router-redux": "^4.0.0",
-    "react-router-relay": "^0.9.0",
-    "redux": "^3.3.1",
-    "redux-thunk": "^1.0.3",
-    "url": "^0.11.0",
-    "yargs": "^4.2.0"
+    "react-router-relay": "^0.9.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.6.4",
-    "babel-core": "^6.6.4",
+    "babel-cli": "^6.6.5",
     "babel-eslint": "^5.0.0",
     "babel-loader": "^6.2.4",
-    "babel-plugin-react-transform": "^2.0.1",
     "babel-plugin-transform-runtime": "^6.6.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.0",
     "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.6.0",
-    "babel-relay-plugin": "^0.7.3",
     "babel-relay-plugin-loader": "^0.7.1",
     "babel-runtime": "^6.6.1",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^5.2.0",
-    "chai-enzyme": "^0.4.1",
-    "cheerio": "^0.20.0",
-    "codecov": "^1.0.1",
     "css-loader": "^0.23.1",
     "cssnano": "^3.5.2",
-    "enzyme": "^2.0.0",
-    "eslint": "2.2.0",
+    "eslint": "2.3.0",
     "eslint-config-standard": "^5.1.0",
     "eslint-config-standard-react": "^2.3.0",
     "eslint-loader": "^1.3.0",
