@@ -1,14 +1,20 @@
 import React, { PropTypes } from 'react'
+import { Link } from 'react-router'
 import PureRenderMixin from 'react-addons-pure-render-mixin'
+import Icon from 'components/Icon/Icon'
 import classes from './ProjectSelection.scss'
+
+const ProjectPropType = React.PropTypes.shape({
+  name: React.PropTypes.string.isRequired,
+  id: React.PropTypes.string.isRequired,
+})
 
 export default class ProjectSelection extends React.Component {
   static propTypes = {
     select: PropTypes.func.isRequired,
-    selectedProject: PropTypes.string.isRequired,
-    projects: PropTypes.arrayOf(React.PropTypes.shape({
-      name: React.PropTypes.string.isRequired,
-    })),
+    add: PropTypes.func.isRequired,
+    selectedProject: ProjectPropType.isRequired,
+    projects: PropTypes.arrayOf(ProjectPropType).isRequired,
   };
 
   constructor (props) {
@@ -16,10 +22,16 @@ export default class ProjectSelection extends React.Component {
 
     this.shouldComponentUpdate = PureRenderMixin.shouldComponentUpdate.bind(this)
     this._onSelect = ::this._onSelect
+    this._toggle = ::this._toggle
+    this._onAdd = ::this._onAdd
 
     this.state = {
       expanded: false,
     }
+  }
+
+  _toggle () {
+    this.setState({ expanded: !this.state.expanded })
   }
 
   _onSelect (e) {
@@ -27,18 +39,43 @@ export default class ProjectSelection extends React.Component {
     this.props.select(projectName)
   }
 
+  _onAdd () {
+    this.props.add()
+  }
+
   render () {
     return (
       <div className={classes.root}>
-        {!this.state.expanded &&
-          <div className={classes.dropdown} />
-        }
+        <div className={classes.head} onClick={this._toggle}>
+          {this.props.selectedProject.name}
+          <div className={`${classes.arrow} ${this.state.expanded ? classes.up : ''}`}>
+            <Icon
+              glyph={require('assets/icons/arrow.svg')}
+              color='#fff'
+              />
+          </div>
+        </div>
         {this.state.expanded &&
-          <select onChange={this._onSelect} value={this.props.selectedProject}>
+          <div className={classes.overlay}>
             {this.props.projects.map((project) => (
-              <option key={project.name} value={project.name}>{project.name}</option>
+              <Link
+                key={project.name}
+                className={classes.listElement}
+                onClick={this._toggle}
+                to={`/${project.id}`}
+                activeClassName={classes.listElementActive}
+                >
+                {project.name}
+                <div title='Duplicate' className={classes.listElementDuplicate}>
+                  <Icon
+                    glyph={require('assets/icons/model.svg')}
+                    color='#fff'
+                    />
+                </div>
+              </Link>
             ))}
-          </select>
+            <div className={classes.add} onClick={this._onAdd}>+ New Project</div>
+          </div>
         }
       </div>
     )
