@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { RelayRouter } from 'react-router-relay'
 import { browserHistory } from 'react-router'
-import routes from './routes'
-import { updateNetworkLayer } from './utils/relay'
+import routes from './routes.js'
+import { updateNetworkLayer } from './utils/relay.js'
 
 updateNetworkLayer()
 
